+++ libusb1-sys/build.rs
@@ -1,6 +1,6 @@
 use std::{env, fs, path::PathBuf};
 
-static VERSION: &str = "1.0.24";
+static VERSION: &str = "1.0.27";
 
 fn link(name: &str, bundled: bool) {
     use std::env::var;
@@ -19,17 +19,6 @@ pub fn link_framework(name: &str) {
     println!("cargo:rustc-link-lib=framework={}", name);
 }
 
-#[cfg(target_env = "msvc")]
-fn find_libusb_pkg(_statik: bool) -> bool {
-    match vcpkg::Config::new().find_package("libusb") {
-        Ok(_) => true,
-        Err(e) => {
-            println!("Can't find libusb pkg: {:?}", e);
-            false
-        }
-    }
-}
-
 fn get_macos_major_version() -> Option<usize> {
     if !cfg!(target_os = "macos") {
         return None;
@@ -45,8 +34,21 @@ fn get_macos_major_version() -> Option<usize> {
     Some(major)
 }
 
-#[cfg(not(target_env = "msvc"))]
 fn find_libusb_pkg(statik: bool) -> bool {
+    if std::env::var("CARGO_CFG_TARGET_ENV") == Ok("msvc".into()) {
+        #[cfg(target_os = "windows")]
+        return match vcpkg::Config::new().find_package("libusb") {
+            Ok(_) => true,
+            Err(e) => {
+                if pkg_config::probe_library("libusb-1.0").is_ok() {
+                    true
+                } else {
+                    println!("Can't find libusb pkg: {:?}", e);
+                    false
+                }
+            }
+        };
+    }
     // https://github.com/rust-lang/rust/issues/96943
     let needs_rustc_issue_96943_workaround: bool = get_macos_major_version()
         .map(|major| major >= 11)
@@ -156,14 +158,16 @@ fn make_source() {
             Some("__attribute__((visibility(\"default\")))"),
         );
 
-        if let Ok(lib) = pkg_config::probe_library("libudev") {
-            base_config.define("USE_UDEV", Some("1"));
-            base_config.define("HAVE_LIBUDEV", Some("1"));
-            base_config.file(libusb_source.join("libusb/os/linux_udev.c"));
-            for path in lib.include_paths {
-                base_config.include(path.to_str().unwrap());
-            }
-        };
+        if std::env::var("CARGO_CFG_TARGET_OS") != Ok("android".into()) {
+            if let Ok(lib) = pkg_config::probe_library("libudev") {
+                base_config.define("USE_UDEV", Some("1"));
+                base_config.define("HAVE_LIBUDEV", Some("1"));
+                base_config.file(libusb_source.join("libusb/os/linux_udev.c"));
+                for path in lib.include_paths {
+                    base_config.include(path.to_str().unwrap());
+                }
+            };
+        }
 
         println!("Including posix!");
         base_config.file(libusb_source.join("libusb/os/events_posix.c"));
@@ -171,8 +175,9 @@ fn make_source() {
     }
 
     if std::env::var("CARGO_CFG_TARGET_OS") == Ok("windows".into()) {
-        #[cfg(target_env = "msvc")]
-        base_config.flag("/source-charset:utf-8");
+        if std::env::var("CARGO_CFG_TARGET_ENV") == Ok("msvc".into()) {
+            base_config.flag("/source-charset:utf-8");
+        }
 
         base_config.warnings(false);
         base_config.define("OS_WINDOWS", Some("1"));